#101 modelで共通な処理をモジュール化
