+++ config/locales/ja.yml
@@ -2,14 +2,14 @@ ja:
   activerecord:
     attributes:
       report:
-        report_date: ""
-        title: ""
-        contents: ""
+        report_date: "日付"
+        title: "タイトル"
+        contents: "内容"
       user:
-        name: ""
-        email: ""
-        password: ""
-        password_confirmation: ""
+        name: "名前"
+        email: "メールアドレス"
+        password: "パスワード"
+        password_confirmation: "パスワード(確認用)"
     errors:
       models:
         report:
